src/requirements.txt
fastapi
uvicorn
python-dotenv
openai
azure-search-documents
PyPDF2
fpdf
requests
